+++ src/button/index.tsx
@@ -46,9 +46,9 @@ export interface ButtonState {
 }
 
 type ButtonActionEvents =
-  React.MouseEvent<HTMLElement> |
-  React.TouchEvent<HTMLElement> |
-  React.FocusEventHandler<HTMLElement>
+  | React.MouseEvent<HTMLElement>
+  | React.TouchEvent<HTMLElement>
+  | React.FocusEventHandler<HTMLElement>
 type functionEvent = (event: ButtonActionEvents) => void
 
 type TypeProps = {
@@ -73,7 +73,7 @@ export const eventHandler = (
   childEvent && childEvent(event)
 }
 
-export default class Button extends PureComponent <ButtonProps, ButtonState> {
+export default class Button extends PureComponent<ButtonProps, ButtonState> {
   private button: HTMLButtonElement
 
   static STATUS = ButtonStatus
@@ -108,12 +108,25 @@ export default class Button extends PureComponent <ButtonProps, ButtonState> {
 
   render() {
     const {
-      children, className, type, href, title,
+      children,
+      className,
+      type,
+      href,
+      title,
       // Modifiers
-      status, icon, shadowed,
+      status,
+      icon,
+      shadowed,
       // Actions
-      onClick, onTouchStart, onTouchEnd, onMouseDown, onMouseUp, onBlur, onFocus,
-      onDoneAnimationEnd, focus,
+      onClick,
+      onTouchStart,
+      onTouchEnd,
+      onMouseDown,
+      onMouseUp,
+      onBlur,
+      onFocus,
+      onDoneAnimationEnd,
+      focus,
       // Extend case of the button for the expand component
       ...attrs
     } = this.props
@@ -150,21 +163,28 @@ export default class Button extends PureComponent <ButtonProps, ButtonState> {
       <Component
         className={cc([
           prefix({ button: true }),
-          prefix({
-            [status]: status, icon: iconSize, shadowed,
-          }, 'kirk-button'),
+          prefix(
+            {
+              [status]: status,
+              icon: iconSize,
+              shadowed,
+            },
+            'kirk-button',
+          ),
           className,
         ])}
         {...typeProps}
         {...attrs}
       >
-        {hasLoader && <Loader
-          size={48}
-          inline
-          done={status === ButtonStatus.CHECKED}
-          onDoneAnimationEnd={onDoneAnimationEnd}
-        />}
-        <span>{children}</span>
+        {hasLoader && (
+          <Loader
+            size={48}
+            inline
+            done={status === ButtonStatus.CHECKED}
+            onDoneAnimationEnd={onDoneAnimationEnd}
+          />
+        )}
+        {children && <span>{children}</span>}
         <style jsx>{style}</style>
       </Component>
     )